rewrited save with simpleJdbcInsert

+++ src/main/kotlin/ru/itmo/coffee/store/repository/AddressRepository.kt
@@ -3,7 +3,7 @@ package ru.itmo.coffee.store.repository
 import ru.itmo.coffee.store.dao.Address
 
 interface AddressRepository {
-    fun save(address: Address): Int
+    fun save(address: Address): Long
     fun update(address: Address): Int
     fun deleteById(id: Long): Int
     fun findAll(): List<Address>
