+++ src/main/kotlin/ru/itmo/coffee/store/repository/CoffeeRepository.kt
@@ -3,7 +3,7 @@ package ru.itmo.coffee.store.repository
 import ru.itmo.coffee.store.dao.Coffee
 
 interface CoffeeRepository {
-    fun save(coffee: Coffee): Int
+    fun save(coffee: Coffee): Long
     fun update(coffee: Coffee): Int
     fun deleteById(id: Long): Int
     fun findAll(): List<Coffee>
